Move Roles enum to types package
