+++ packages/backend-core/src/users/utils.ts
@@ -1,4 +1,10 @@
-import { CloudAccount, ContextUser, User, UserGroup } from "@budibase/types"
+import {
+  CloudAccount,
+  ContextUser,
+  User,
+  UserGroup,
+  Roles as BUILTIN_ROLE_IDS,
+} from "@budibase/types"
 import * as accountSdk from "../accounts"
 import env from "../environment"
 import { getPlatformUser } from "./lookup"
@@ -6,7 +12,6 @@ import { EmailUnavailableError } from "../errors"
 import { getTenantId } from "../context"
 import { sdk } from "@budibase/shared-core"
 import { getAccountByTenantId } from "../accounts"
-import { BUILTIN_ROLE_IDS } from "../security/roles"
 import * as context from "../context"
 
 // extract from shared-core to make easily accessible from backend-core
