+++ packages/client/src/stores/derived/currentRole.js
@@ -2,7 +2,7 @@ import { derived } from "svelte/store"
 import { devToolsStore } from "../devTools.js"
 import { authStore } from "../auth.js"
 import { devToolsEnabled } from "./devToolsEnabled.js"
-import { Roles } from "@budibase/shared-core"
+import { Roles } from "@budibase/types"
 
 // Derive the current role of the logged-in user
 export const currentRole = derived(
