+++ packages/shared-core/src/constants/index.ts
@@ -179,13 +179,3 @@ export enum BpmStatusValue {
   VERIFYING_EMAIL = "verifying_email",
   COMPLETED = "completed",
 }
-
-// Role IDs
-export enum Roles {
-  ADMIN = "ADMIN",
-  POWER = "POWER",
-  BASIC = "BASIC",
-  PUBLIC = "PUBLIC",
-  BUILDER = "BUILDER",
-  CREATOR = "CREATOR",
-}
