+++ packages/shared-core/src/sdk/documents/users.ts
@@ -4,10 +4,10 @@ import {
   SEPARATOR,
   User,
   InternalTable,
+  Roles,
 } from "@budibase/types"
 import { getProdAppID } from "./applications"
 import * as _ from "lodash/fp"
-import { Roles } from "../../constants/index"
 
 // checks if a user is specifically a builder, given an app ID
 export function isBuilder(user: User | ContextUser, appId?: string): boolean {
