+++ packages/types/src/sdk/index.ts
@@ -22,3 +22,4 @@ export * from "./permissions"
 export * from "./row"
 export * from "./vm"
 export * from "./view"
+export * from "./roles"
