+++ packages/types/src/sdk/roles.ts
@@ -1,0 +1,9 @@
+// Role IDs
+export enum Roles {
+  ADMIN = "ADMIN",
+  POWER = "POWER",
+  BASIC = "BASIC",
+  PUBLIC = "PUBLIC",
+  BUILDER = "BUILDER",
+  CREATOR = "CREATOR",
+}
